Seperated unit tests and tested working with Binja 3.3
