+++ README.md
@@ -6,7 +6,8 @@ _Android Dex disassembler and Binary Ninja plugin_
 
 ## Description
 
-Banjo parses Dex files and disassembles them into a smali syntax that is close to [baksmali](https://github.com/JesusFreke/smali)'s.
+Banjo parses Dex files and disassembles them into a smali syntax that is close
+to [baksmali](https://github.com/JesusFreke/smali)'s.
 
 There are three parts to this project:
 
@@ -18,7 +19,8 @@ For more documentation, see the [doc/](doc) directory.
 
 There are still some rough edges. See [GitHub issues](https://github.com/CarveSystems/banjo/issues) for more details.
 
-This project was released at a [ShmooCon 2020 talk](https://github.com/CarveSystems/presentations/tree/master/2020/banjo).
+This project was released at
+a [ShmooCon 2020 talk](https://github.com/CarveSystems/presentations/tree/master/2020/banjo).
 
 ## Installation Instructions
