+++ plugin.json
@@ -1,18 +1,26 @@
 {
-    "pluginmetadataversion": 2,
-    "name": "Banjo",
-    "author": "Austin Ralls",
-    "type": ["binaryview", "architecture"],
-    "api": ["python3"],
-    "description": "Adds support for Android Dex files",
-    "longdescription": "Banjo parses Dex files and disassembles them into a smali syntax that is close to [baksmali](https://github.com/JesusFreke/smali)'s.\n\nThe plugin still has rough edges. See GitHub issues for more details.",
-	",
-    "license": {
-        "name": "MIT",
-        "text": "Copyright 2020 Austin Ralls\n\nPermission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the \"Software\"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:\n\nThe above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.\n\nTHE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE."
-    },
-    "platforms": ["Linux"],
-    "installinstructions": {"Linux": "Make sure Binary Ninja is using Python 3.8"},
-    "version": "1.0.0",
-    "minimumbinaryninjaversion": 2003
+  "pluginmetadataversion": 2,
+  "name": "Banjo",
+  "author": "Austin Ralls",
+  "type": [
+    "binaryview",
+    "architecture"
+  ],
+  "api": [
+    "python3"
+  ],
+  "description": "Adds support for Android Dex files",
+  "longdescription": "Banjo parses Dex files and disassembles them into a smali syntax that is close to [baksmali](https://github.com/JesusFreke/smali)'s.\n\nThe plugin still has rough edges. See GitHub issues for more details.",
+  "license": {
+    "name": "MIT",
+    "text": "Copyright 2020 Austin Ralls\n\nPermission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the \"Software\"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions:\n\nThe above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.\n\nTHE SOFTWARE IS PROVIDED \"AS IS\", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE."
+  },
+  "platforms": [
+    "Linux"
+  ],
+  "installinstructions": {
+    "Linux": "Make sure Binary Ninja is using Python 3.8"
+  },
+  "version": "1.0.0",
+  "minimumbinaryninjaversion": 2003
 }
